Retry NoMessageException in DataPartsExchange
If a connection from connection pool is idle for too long the server
will try to close it, leaving the socket in CLOSE-WAIT state. As there
is no connection manager that will poll connections to check if they are
still alive there is a chance that we'll get a connection in this state
out of the pool. Writing to a socket in CLOSE-WAIT can succeed, and
Poco's test for keep alive timeout doesn't seem to work correctly.

Solution: retry this error until we succeed. Poco will create new
connections for us.
Alternative solution: Try to peek one byte from connection before
returning it from the pool, may look cleaner, but there is still a very
low chance of connection getting closed by the time we try to read from
it.

strace example:

```
218921 sendto(546, "POST /?endpoint=DataPartsExchang"..., 207, 0, NULL, 0 <unfinished ...>
218921 <... sendto resumed> )           = 207
218921 recvfrom(546,  <unfinished ...>
218921 <... recvfrom resumed> "", 4096, 0, NULL, NULL) = 0
218921 close(546 <unfinished ...>
218921 <... close resumed> )            = 0
218921 write(2, "2019.03.28 18:01:54.082307 [ 27 "..., 315 <unfinished ...>
```

and log message:

```
2019.03.28 18:01:54.082307 [ 27 ] {} <Error> table_name (StorageReplicatedMergeTree): DB::StorageReplicatedMergeTree::queueTask()::<lambda(DB::StorageReplicatedMergeTree::LogEntryPtr&)>: Poco::Exception. Code: 1000, e.code() = 0, e.displayText() = No message received, e.what() = No message received
```

Closes 4047.